Split off definition file processing into separate service
